Datamodel: Rename interface Checkpoint to ICheckpoint
We will only have a single implementation of all classes in the datamodel
so there is no good pre-/suffix to distinguish the implementations via
their class names and thus the interfaces must be distinguished instead.
I.e. the class implementing this interface will be called "Checkpoint"
because there is no reasonable different name.