Datamodel: Rename abstract class Timestamps to ITimestamps
See commit 883d377b863a6219857b2dfe4d765c3d3a365390 for the reason behind
renaming.

Using the interface prefix even though it is abstract because it is only
not an interface yet because we need to provide a final implementation for
one function for now due to the reasons explained there.