JavaSHA256Test: Fix testEquals() to ensure equals() is called upon the class at test and not the one we want to pass objects of as argument to equals()
This is mostly subject to the last two modified ones, but I'm changing
all others anyway to keep the pattern coherent.

I have not checked the JUnit JavaDoc for which of the arguments of
assertEquals() it calls equals(), but even if it were the right one it is
probably a safer programming style to not assume this.