ConcurrentCheckpointGenerator: Set name for worker threads to ease debugging
There are 4 threads visible in htop currently for me which create 100%
CPU usage for many minutes on a core even though Checkpoint only does ~ 5 MiB/s
on the rotational disk I am testing it at.

This is very suspicious because if they were eligible to run for so long
with so high CPU usage they should be dealing with a large file which
they're hashing.
But then the disk wouldn't be doing only 5 MiB/s.

So in order to debug that it would be beneficial to figure out if those threads
are even worker threads or something else.