CreateCommand: Adapt to ability of ConcurrentCheckpointGenerator to use a different default thread count for SSDs
And move printing of arguments we've parsed to ConcurrentCheckpoitGenerator
to ensure it prints what's actually being used.
This is a good idea since this very same commit shows that it would be
easy to break the code which decides the default by not realizing that
the constructor of ConcurrentCheckpointGenerator consumes an Integer
for the thread count so it can be null to allow for the default to be
used.