NodeFinder: Make related error messages more distinguished among each other
The way they were before made it look like each file which is ignored
as part of a different filesystem would be listed, but that is not the
case:
We only list excluded files which are inside a dir on the main FS but
on a different filesystems individually.
Files whose parent dir is on a different FS are not listed.