Checkpoint.save(): Remove totally overkill micro optimization
The readability aspect of having the structure of the code match
the one of the code for saving the other timestamps right below
is much more important.
And the fact that btime cannot be obtained by Linux' userland yet
is an implementation detail which isn't relevant to the serialization
code here.